POS Tagger using LSTM Sequnce Modeling
Parts of Speech Tagger using Bidirectional LSTM Sequence.
Dataset: ner_dataset which is a named entity recognition dataset.words and tags are converted to numbersword2id and tag2id are created.word as input and tag as output are preparedoutput is converted to one-hot categorical matrixModel with Embedding, Bidirectional LSTM and TimeDistributed Layers is created.Model is compiled with rmsprop opitimizerPrediction is done with one of the test data and a custom sentence.